create initial channel in the UI
In many cases, it is preferable for the caller to create the first
channel. For example, `spin deploy` creates a channel called
`spin-deploy` with a different revision selection strategy than what the
UI would create.

To accomodate different workflows, I moved the initial channel creation
from the API into the UI.

Signed-off-by: Matthew Fisher <matt.fisher@fermyon.com>